@@ -8,13 +8,14 @@
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
 	"bin": {
-		"has-yarn": "cli.js"
+		"has-yarn": "./cli.js"
 	},
 	"engines": {
-		"node": ">=8"
+		"node": "^12.20.0 || ^14.13.1 || >=16.0.0"
 	},
 	"scripts": {
 		"test": "xo && ava"
@@ -37,13 +38,13 @@
 		"npm"
 	],
 	"dependencies": {
-		"has-yarn": "^1.0.0",
-		"log-symbols": "^2.2.0",
-		"meow": "^5.0.0"
+		"has-yarn": "^3.0.0",
+		"log-symbols": "^5.0.0",
+		"meow": "^10.1.1"
 	},
 	"devDependencies": {
-		"ava": "^1.2.1",
-		"execa": "^1.0.0",
-		"xo": "^0.24.0"
+		"ava": "^3.15.0",
+		"execa": "^5.1.1",
+		"xo": "^0.44.0"
 	}
 }
